Move parseDate.js tests to their own file
