+++ root/static/scripts/tests/utility.js
@@ -20,7 +20,6 @@ import compareDates, {
 import formatDate from '../common/utility/formatDate.js';
 import formatDatePeriod from '../common/utility/formatDatePeriod.js';
 import formatSetlist from '../common/utility/formatSetlist.js';
-import parseDate from '../common/utility/parseDate.js';
 import * as fullwidthLatin from '../edit/utility/fullwidthLatin.js';
 import isShortenedUrl from '../edit/utility/isShortenedUrl.js';
 
@@ -164,36 +163,6 @@ test('compareDatePeriods', function (t) {
     {begin_date: null, end_date: null, ended: true},
     {begin_date: null, end_date: null, ended: false},
   ) < 0, 'ended date periods are sorted before non-ended ones');
-});
-
-test('parseDate', function (t) {
-  t.plan(16);
-
-  var parseDateTests = [
-    {date: '', expected: {year: null, month: null, day: null}},
-    {date: '0000', expected: {year: 0, month: null, day: null}},
-    {date: '1999-01-02', expected: {year: 1999, month: 1, day: 2}},
-    {date: '1999-01', expected: {year: 1999, month: 1, day: null}},
-    {date: '1999', expected: {year: 1999, month: null, day: null}},
-    {date: '????-01-02', expected: {year: null, month: 1, day: 2}},
-    {date: '????-??-02', expected: {year: null, month: null, day: 2}},
-    {date: '1999-??-02', expected: {year: 1999, month: null, day: 2}},
-
-    // Relationship editor seeding format (via URL query params).
-    {date: '-----', expected: {year: null, month: null, day: null}},
-    {date: '----02', expected: {year: null, month: null, day: 2}},
-    {date: '--01--', expected: {year: null, month: 1, day: null}},
-    {date: '--01-02', expected: {year: null, month: 1, day: 2}},
-    {date: '1999--', expected: {year: 1999, month: null, day: null}},
-    {date: '1999----', expected: {year: 1999, month: null, day: null}},
-    {date: '1999---02', expected: {year: 1999, month: null, day: 2}},
-    {date: '1999-01--', expected: {year: 1999, month: 1, day: null}},
-  ];
-
-  for (const test of parseDateTests) {
-    const result = parseDate(test.date);
-    t.deepEqual(result, test.expected, test.date);
-  }
 });
 
 test('formatDate', function (t) {
